# Precise tag benchmarks showing improvement from haskell-tree-sitter optimizations