--- a/bench/Main.hs
+++ b/bench/Main.hs
@@ -1,9 +1,12 @@
+{-# OPTIONS_GHC -Wno-unused-imports -Wno-unused-top-binds #-}
 module Main (main) where
 
 import Gauge
 import qualified Evaluation
+import qualified Tagging
 
 main :: IO ()
 main = defaultMain
-  [ Evaluation.benchmarks
+  [ Tagging.benchmarks
+  -- , Evaluation.benchmarks
   ]