--- a/bench/Main.hs
+++ b/bench/Main.hs
@@ -1,4 +1,3 @@
-{-# OPTIONS_GHC -Wno-unused-imports -Wno-unused-top-binds #-}
 module Main (main) where
 
 import Gauge