--- a/bench/Tagging.hs
+++ b/bench/Tagging.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards   #-}
 {-# LANGUAGE TypeApplications  #-}
 
 {-# OPTIONS_GHC -Wno-unused-imports -Wno-unused-top-binds #-}