--- a/cabal.project
+++ b/cabal.project
@@ -10,7 +10,6 @@ packages: .
           semantic-tsx
           semantic-typescript
           semantic-tags
-          ../haskell-tree-sitter/tree-sitter*
 
 jobs: $ncpus