--- a/semantic-go/src/Language/Go.hs
+++ b/semantic-go/src/Language/Go.hs
@@ -13,8 +13,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: Go.SourceFile a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . GoTags.tags . getTerm