--- a/semantic-go/src/Language/Go.hs
+++ b/semantic-go/src/Language/Go.hs
@@ -5,6 +5,7 @@ module Language.Go
 ) where
 
 
+import           Data.Proxy
 import qualified Language.Go.Tags as GoTags
 import qualified Tags.Tagging.Precise as Tags
 import qualified TreeSitter.Go (tree_sitter_go)
@@ -14,10 +15,11 @@ import qualified TreeSitter.Unmarshal as TS
 newtype Term a = Term { getTerm :: Go.SourceFile a }
 
 instance TS.SymbolMatching Term where
-  showFailure _ _ = "failed for Term"
+  matchedSymbols _ = TS.matchedSymbols (Proxy :: Proxy Go.SourceFile)
+  showFailure _ = TS.showFailure (Proxy :: Proxy Go.SourceFile)
 
 instance TS.Unmarshal Term where
-  matchers = fmap (TS.hoist Term) TS.matchers
+  matchers = fmap (fmap (TS.hoist Term)) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . GoTags.tags . getTerm