--- a/semantic-java/semantic-java.cabal
+++ b/semantic-java/semantic-java.cabal
@@ -27,8 +27,8 @@ library
     , fused-effects    ^>= 1.0
     , semantic-source  ^>= 0.0
     , semantic-tags    ^>= 0.0
-    , tree-sitter      ^>= 0.7.2
-    , tree-sitter-java ^>= 0.6
+    , tree-sitter      ^>= 0.8
+    , tree-sitter-java ^>= 0.6.1
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options: