--- a/semantic-java/src/Language/Java.hs
+++ b/semantic-java/src/Language/Java.hs
@@ -12,8 +12,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: Java.Program a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . JavaTags.tags . getTerm