--- a/semantic-json/semantic-json.cabal
+++ b/semantic-json/semantic-json.cabal
@@ -24,8 +24,8 @@ library
   build-depends:
       base              >= 4.13 && < 5
     , semantic-tags    ^>= 0.0
-    , tree-sitter      ^>= 0.7.2
-    , tree-sitter-json ^>= 0.5
+    , tree-sitter      ^>= 0.8
+    , tree-sitter-json ^>= 0.6
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options: