--- a/semantic-json/src/Language/JSON.hs
+++ b/semantic-json/src/Language/JSON.hs
@@ -4,6 +4,7 @@ module Language.JSON
 , TreeSitter.JSON.tree_sitter_json
 ) where
 
+import           Data.Proxy
 import qualified Tags.Tagging.Precise as Tags
 import qualified TreeSitter.JSON (tree_sitter_json)
 import qualified TreeSitter.JSON.AST as JSON
@@ -12,10 +13,11 @@ import qualified TreeSitter.Unmarshal as TS
 newtype Term a = Term { getTerm :: JSON.Document a }
 
 instance TS.SymbolMatching Term where
-  showFailure _ _ = "failed for Term"
+  matchedSymbols _ = TS.matchedSymbols (Proxy :: Proxy JSON.Document)
+  showFailure _ = TS.showFailure (Proxy :: Proxy JSON.Document)
 
 instance TS.Unmarshal Term where
-  matchers = fmap (TS.hoist Term) TS.matchers
+  matchers = fmap (fmap (TS.hoist Term)) TS.matchers
 
 -- | Tags aren’t really meaningful for JSON, but by implementing this we can avoid having to customize the set of parsers used for computing tags.
 instance Tags.ToTags Term where