--- a/semantic-python/semantic-python.cabal
+++ b/semantic-python/semantic-python.cabal
@@ -28,8 +28,8 @@ common haskell
                      , semantic-source ^>= 0.0
                      , semantic-tags ^>= 0.0
                      , text ^>= 1.2.3
-                     , tree-sitter ^>= 0.7.2
-                     , tree-sitter-python ^>= 0.8
+                     , tree-sitter ^>= 0.8
+                     , tree-sitter-python ^>= 0.8.1
 
   ghc-options:
     -Weverything