--- a/semantic-python/src/Language/Python.hs
+++ b/semantic-python/src/Language/Python.hs
@@ -12,8 +12,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: Py.Module a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . PyTags.tags . getTerm