--- a/semantic-ruby/src/Language/Ruby.hs
+++ b/semantic-ruby/src/Language/Ruby.hs
@@ -17,8 +17,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: Rb.Program a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . evalState @[Text] [] . PyTags.tags . getTerm