--- a/semantic-ruby/src/Language/Ruby.hs
+++ b/semantic-ruby/src/Language/Ruby.hs
@@ -6,10 +6,10 @@ module Language.Ruby
 , TreeSitter.Ruby.tree_sitter_ruby
 ) where
 
-
-import Control.Carrier.State.Strict
-import Data.Text (Text)
-import qualified Language.Ruby.Tags as PyTags
+import           Control.Carrier.State.Strict
+import           Data.Proxy
+import           Data.Text (Text)
+import qualified Language.Ruby.Tags as RbTags
 import qualified Tags.Tagging.Precise as Tags
 import qualified TreeSitter.Ruby (tree_sitter_ruby)
 import qualified TreeSitter.Ruby.AST as Rb
@@ -18,10 +18,11 @@ import qualified TreeSitter.Unmarshal as TS
 newtype Term a = Term { getTerm :: Rb.Program a }
 
 instance TS.SymbolMatching Term where
-  showFailure _ _ = "failed for Term"
+  matchedSymbols _ = TS.matchedSymbols (Proxy :: Proxy Rb.Program)
+  showFailure _ = TS.showFailure (Proxy :: Proxy Rb.Program)
 
 instance TS.Unmarshal Term where
-  matchers = fmap (TS.hoist Term) TS.matchers
+  matchers = fmap (fmap (TS.hoist Term)) TS.matchers
 
 instance Tags.ToTags Term where
-  tags src = Tags.runTagging src . evalState @[Text] [] . PyTags.tags . getTerm
+  tags src = Tags.runTagging src . evalState @[Text] [] . RbTags.tags . getTerm