--- a/semantic-tsx/semantic-tsx.cabal
+++ b/semantic-tsx/semantic-tsx.cabal
@@ -28,8 +28,8 @@ common haskell
                      , semantic-source ^>= 0.0
                      , semantic-tags ^>= 0.0
                      , text ^>= 1.2.3
-                     , tree-sitter ^>= 0.7.2
-                     , tree-sitter-tsx ^>= 0.4
+                     , tree-sitter ^>= 0.8
+                     , tree-sitter-tsx ^>= 0.4.1
 
   ghc-options:
     -Weverything