--- a/semantic-tsx/src/Language/TSX.hs
+++ b/semantic-tsx/src/Language/TSX.hs
@@ -14,8 +14,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: TSX.Program a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . TsxTags.tags . getTerm