--- a/semantic-tsx/src/Language/TSX.hs
+++ b/semantic-tsx/src/Language/TSX.hs
@@ -5,7 +5,7 @@ module Language.TSX
 , TreeSitter.TSX.tree_sitter_tsx
 ) where
 
-
+import           Data.Proxy
 import qualified Language.TSX.Tags as TsxTags
 import qualified Tags.Tagging.Precise as Tags
 import qualified TreeSitter.TSX (tree_sitter_tsx)
@@ -15,10 +15,11 @@ import qualified TreeSitter.Unmarshal as TS
 newtype Term a = Term { getTerm :: TSX.Program a }
 
 instance TS.SymbolMatching Term where
-  showFailure _ _ = "failed for Term"
+  matchedSymbols _ = TS.matchedSymbols (Proxy :: Proxy TSX.Program)
+  showFailure _ = TS.showFailure (Proxy :: Proxy TSX.Program)
 
 instance TS.Unmarshal Term where
-  matchers = fmap (TS.hoist Term) TS.matchers
+  matchers = fmap (fmap (TS.hoist Term)) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . TsxTags.tags . getTerm