--- a/semantic-typescript/src/Language/TypeScript.hs
+++ b/semantic-typescript/src/Language/TypeScript.hs
@@ -14,8 +14,11 @@ import qualified TreeSitter.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: TypeScript.Program a }
 
+instance TS.SymbolMatching Term where
+  showFailure _ _ = "failed for Term"
+
 instance TS.Unmarshal Term where
-  unmarshalNode node = Term <$> TS.unmarshalNode node
+  matchers = fmap (TS.hoist Term) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . TsTags.tags . getTerm