--- a/semantic-typescript/src/Language/TypeScript.hs
+++ b/semantic-typescript/src/Language/TypeScript.hs
@@ -5,7 +5,7 @@ module Language.TypeScript
 , TreeSitter.TypeScript.tree_sitter_typescript
 ) where
 
-
+import           Data.Proxy
 import qualified Language.TypeScript.Tags as TsTags
 import qualified Tags.Tagging.Precise as Tags
 import qualified TreeSitter.TypeScript (tree_sitter_typescript)
@@ -15,10 +15,11 @@ import qualified TreeSitter.Unmarshal as TS
 newtype Term a = Term { getTerm :: TypeScript.Program a }
 
 instance TS.SymbolMatching Term where
-  showFailure _ _ = "failed for Term"
+  matchedSymbols _ = TS.matchedSymbols (Proxy :: Proxy TypeScript.Program)
+  showFailure _ = TS.showFailure (Proxy :: Proxy TypeScript.Program)
 
 instance TS.Unmarshal Term where
-  matchers = fmap (TS.hoist Term) TS.matchers
+  matchers = fmap (fmap (TS.hoist Term)) TS.matchers
 
 instance Tags.ToTags Term where
   tags src = Tags.runTagging src . TsTags.tags . getTerm