--- a/src/Parsing/TreeSitter.hs
+++ b/src/Parsing/TreeSitter.hs
@@ -9,7 +9,6 @@ module Parsing.TreeSitter
 
 import Prologue
 
-import           Control.Carrier.Fail.Either
 import           Control.Carrier.Reader
 import qualified Control.Exception as Exc
 import           Foreign
@@ -58,8 +57,8 @@ parseToPreciseAST
   -> m (Either TSParseException (t Loc))
 parseToPreciseAST parseTimeout language blob = runParse parseTimeout language blob $ \ rootPtr ->
   TS.withCursor (castPtr rootPtr) $ \ cursor ->
-    runFail (runReader cursor (runReader (Source.bytes (blobSource blob)) (TS.peekNode >>= TS.unmarshalNode)))
-      >>= either (Exc.throw . UnmarshalFailure) pure
+    (Right <$> runReader (TS.UnmarshalState (Source.bytes (blobSource blob)) cursor) (TS.peekNode >>= TS.unmarshalNode))
+      >>= either (Exc.throw . UnmarshalFailure . TS.getUnmarshalError) pure
 
 instance Exception TSParseException where
   displayException = \case