--- a/src/Parsing/TreeSitter.hs
+++ b/src/Parsing/TreeSitter.hs
@@ -57,7 +57,7 @@ parseToPreciseAST
   -> m (Either TSParseException (t Loc))
 parseToPreciseAST parseTimeout language blob = runParse parseTimeout language blob $ \ rootPtr ->
   TS.withCursor (castPtr rootPtr) $ \ cursor ->
-    (Right <$> runReader (TS.UnmarshalState (Source.bytes (blobSource blob)) cursor) (TS.unmarshal cursor))
+    (Right <$> runReader (TS.UnmarshalState (Source.bytes (blobSource blob)) cursor) (liftIO (peek rootPtr) >>= TS.unmarshalNode))
       >>= either (Exc.throw . UnmarshalFailure . TS.getUnmarshalError) pure
 
 instance Exception TSParseException where