… test/fixtures/json/corpus/hash.parseA.txt
… test/fixtures/json/corpus/hash.parseB.txt
… test/fixtures/json/corpus/string.parseA.txt
… test/fixtures/json/corpus/string.parseB.txt