Adds benchmarks and demonstrates impact of optimizations implemented in https://github.com/tree-sitter/haskell-tree-sitter/pull/257.

```
benchmarked tagging/python/precise
time                 51.05 ms   (48.13 ms .. 52.87 ms)
                     0.995 R²   (0.990 R² .. 0.998 R²)
mean                 50.71 ms   (49.92 ms .. 52.28 ms)
std dev              2.143 ms   (1.245 ms .. 2.837 ms)
variance introduced by outliers: 13% (moderately inflated)

benchmarked tagging/python/a la carte
time                 110.0 ms   (108.2 ms .. 112.0 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 109.7 ms   (108.1 ms .. 111.1 ms)
std dev              2.533 ms   (1.673 ms .. 3.618 ms)

benchmarking tagging/go/precise ... took 7.994 s, total 56 iterations
benchmarked tagging/go/precise
time                 142.4 ms   (140.1 ms .. 144.3 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 143.3 ms   (142.3 ms .. 144.7 ms)
std dev              2.052 ms   (1.304 ms .. 2.894 ms)

benchmarking tagging/go/a la carte ... took 22.66 s, total 56 iterations
benchmarked tagging/go/a la carte
time                 409.7 ms   (405.5 ms .. 420.5 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 410.0 ms   (407.9 ms .. 414.0 ms)
std dev              4.758 ms   (2.808 ms .. 7.061 ms)

benchmarked tagging/ruby/precise
time                 14.17 ms   (14.03 ms .. 14.43 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 14.25 ms   (14.18 ms .. 14.33 ms)
std dev              210.9 μs   (168.0 μs .. 285.3 μs)

benchmarked tagging/ruby/a la carte
time                 25.54 ms   (25.06 ms .. 26.17 ms)
                     0.997 R²   (0.993 R² .. 0.999 R²)
mean                 25.50 ms   (25.07 ms .. 25.97 ms)
std dev              1.017 ms   (762.3 μs .. 1.422 ms)
variance introduced by outliers: 10% (moderately inflated)
```

The performance of precise ASTs is now in a good place (improves on a la carte considerably).